automations

+++ orun/core/serializers/jinja2.py
@@ -2,6 +2,7 @@
 Execute mako template before deserialize.
 """
 import os
+import warnings
 from pathlib import Path
 import tempfile
 from jinja2 import Template
@@ -21,14 +22,17 @@ class Deserializer(base.Deserializer):
             self.stream_or_string = self.options['filename'] = filename = Path(filename % {'db_vendor': connection.vendor})
         else:
             filename = Path(filename)
-        s = Template(self.stream.read()).render(deserializer=self)
-        filename = filename.name.rsplit('.', 1)[0]
-        fmt = filename.rsplit('.', 1)[1]
-        deserializer = get_deserializer(fmt)
-        with tempfile.NamedTemporaryFile('r+', suffix=f'.{fmt}', delete=False) as f:
-            try:
-                f.write(s)
-                f.close()
-                deserializer(Path(f.name), self.addon, **self.options).deserialize()
-            finally:
-                os.remove(f.name)
+        if not os.path.isfile(filename):
+            warnings.warn(f'File "{filename}" does not exist')
+        else:
+            s = Template(self.stream.read()).render(deserializer=self)
+            filename = filename.name.rsplit('.', 1)[0]
+            fmt = filename.rsplit('.', 1)[1]
+            deserializer = get_deserializer(fmt)
+            with tempfile.NamedTemporaryFile('r+', suffix=f'.{fmt}', delete=False) as f:
+                try:
+                    f.write(s)
+                    f.close()
+                    deserializer(Path(f.name), self.addon, **self.options).deserialize()
+                finally:
+                    os.remove(f.name)
